-Wno-stringop-overflow vs Warray-bounds
The bogus Wstringop-overflow warnings conditionally issued for
Warray-bounds-48.c and -Wzero-length-array-bounds-2.c are expected
under conditions that depend on the availability of certain vector
patterns, but that don't take cost analysis into account, which
leads to omitted expected warnings and to unexpected warnings.

On riscv64-elf and arm-eabi/-mcpu=cortex-r5, for example, though the
Warray-bounds-48.c condition passes, we don't issue warnings because
we decide not to vectorize the assignments.

On riscv64, for Wzero-length-array-bounds-2.c, we issue the expected
warning in test_C_global_buf, but we also issue a warning for
test_C_local_buf under the same conditions, that would be expected on
other platforms but that is not issued on them.  On
arm-eabi/-mcpu=cortex-r5, the condition passes so we'd expect the
warning in both functions, but we don't warn on either.

Instead of further extending the effective target tests, introduced to
temporarily tolerate these expected bogus warnings, so as to capture
the cost analyses that lead to the mismatched decisions, I'm disabling
the undesired warnings for these two tests.

Regstrapping on x86_64-linux-gnu, also tested with crosses to
riscv64-elf and arm-eabi.  Ok to install?

for  gcc/testsuite/ChangeLog

	PR tree-optimization/102706
	* gcc.dg/Warray-bounds-48.c: Disable -Wstringop-overflow.
	* gcc.dg/Wzero-length-array-bounds-2.c: Likewise.