libstdc++, Darwin: Limit recursive mutex init to OS versions needing it.
While looking at Darwin's os_defines, I wondered whether
_GTHREAD_USE_RECURSIVE_MUTEX_INIT_FUNC was really still necessary. On checking
the OS, it seems that it was a transient problem that only appears in Darwin11.
Earlier versions do not have the initializer, and the operation is fixed in
Darwin12+.  So, I figure we still need the workaround for Darwin8-11 but not
later versions (patch implements this).

Testing on both earlier and later systems does not (as one might expect) reveal
any differences (we still have some lock-related tests that 'twinkle' so that
there can be something still hidden by that).

Anyway. the question is. "is there any reason not to use the system version of
this where it works, or should I just leave the status quo?"  If the former,
then "ok for master?".

thanks
Iain

-- >8 --

The problem described in pr 51906 was fixed in the next OS release.  Limit the
workaround to systems that need it.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>

libstdc++-v3/ChangeLog:

	* config/os/bsd/darwin/os_defines.h
	(_GTHREAD_USE_RECURSIVE_MUTEX_INIT_FUNC): Limit use of this macro
	to OS versions that need it.