c++, driver: Fix -static-libstdc++ for targets without Bstatic/dynamic.
This fixes a long-standing problem on Darwin where we cannot independently set
-static-libstdc++ because the flag gets stripped by the g++ driver.

This patch is essentially the same as the one used for the 'D' driver and has
been in local use for some time.  It has also been tested on Linux.

OK for master?
backports?
thanks
Iain

-- >8 --

The current implementation for swapping between the static and shared c++
runtimes relies on the static linker supporting Bstatic/dynamic which is
not available for every target (Darwin's linker does not support this).

Specs substitution (%s) is an alternative solution for this (which is what
Darwin uses for Fortran, D and Objective-C).  However, specs substitution
requires that the '-static-libstdc++' be preserved in the driver's command
line.  The patch here arranges for this to be done when the configuration
determines that linker support for Bstatic/dynamic is missing.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>

gcc/cp/ChangeLog:

	* g++spec.cc (lang_specific_driver): Preserve -static-libstdc++ in
	the driver command line for targets without -Bstatic/dynamic support
	in their static linker.