tree-optimization/104475 - bogus -Wstringop-overflow
The following avoids a bogus -Wstringop-overflow diagnostic by
properly recognizing that &d->m_mutex cannot be nullptr in C++
even if m_mutex is at offset zero.  The frontend already diagnoses
a &d->m_mutex != nullptr comparison and the following transfers
this knowledge to the middle-end which sees &d->m_mutex as
simple pointer arithmetic.  The new ADDR_NONZERO flag on an
ADDR_EXPR is used to carry this information and it's checked in
the tree_expr_nonzero_p API which causes this to be folded early.

To avoid the bogus diagnostic this avoids separating the nullptr
path via jump-threading by eliminating the nullptr check.

I'd appreciate C++ folks picking this up and put the flag on
the appropriate ADDR_EXPRs - I've tried avoiding to put it on
all of them and didn't try hard to mimick what -Waddress warns
on (the code is big, maybe some refactoring would help but also
not sure what exactly the C++ standard constraints are here).

Bootstrapped and tested on x86_64-unknown-linux-gnu.

Thanks,
Richard.

	PR tree-optimization/104475
gcc/
	* tree-core.h: Document use of nothrow_flag on ADDR_EXPR.
	* tree.h (ADDR_NONZERO): New.
	* fold-const.cc (tree_single_nonzero_warnv_p): Check
	ADDR_NONZERO.

gcc/cp/
	* typeck.cc (cp_build_addr_expr_1): Set ADDR_NONZERO
	on the built address if it is of a COMPONENT_REF.

	* g++.dg/opt/pr104475.C: New testcase.