LoongArch: drop loongarch-driver
Currently the loongarch-driver code tries to parse the values of some -m
flags to numeric representation, and then output the strings
corresponding to the numeric representations. This is mostly an no-op,
and it leads to duplication of these flags' parse code, which makes
building libgccjit fail by having multiple copies of this code to link.
However, this driver code is previously used to generate default values
for these -m flags, which is now obsoleted by respecting the values of
--with-* build-time configure flags (or their default values determined
in config.gcc).

Some other architectures include driver code to handle
-m{arch,tune}=native by converting it to the current CPU type. However
as the LoongArch compiler itself can handle native value, such a driver
is not necessary on LoongArch.

Drop loongarch-driver now as it's mostly useless. The specs snippets
that handles -mabi values are moved to loongarch.h as they're still
useful and does not depend on the driver.

gcc/ChangeLog:

	* config/loongarch/loongarch.h (ABI_GRLEN_SPEC):
	Moved from loongarch-driver.h.
	* config/loongarch/loongarch.h (ABI_SPEC): Ditto.
	* config/loongarch/loongarch-driver.h: Removed.
	* config/loongarch/loongarch-driver.cc: Removed.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>