LoongArch: Fixed a compilation failure with '%c' in inline assembly [PR107731].
There is description of '%c' "%n" "%a" and "%l" in section 17.5 of gccint.pdf.
So I can understand that these descriptors are the ones that the common code
implementation back end has to support, right?
But I don't see the use of these descriptors in gcc.pdf.Now I want to add the
descriptor information under loongarch. I had to add '%c' to the schema section.
Is there a better solution to this?

V2 -> v3:
1. Correct a clerical error.
2. Adding document for loongarch operand modifiers.