AArch64: Enable TARGET_CONST_ANCHOR
Enable TARGET_CONST_ANCHOR to allow complex constants to be created via immediate add.
Use a 24-bit range as that enables a 3 or 4-instruction immediate to be replaced by
2 additions.  Fix the costing of immediate add to support 24-bit immediate and 12-bit shifted
immediates.  The generated code for the testcase is now the same or better than LLVM.
It also results in a small codesize reduction on SPEC.

Passes bootstrap and regress, OK for commit?

gcc/
	* config/aarch64/aarch64.cc (aarch64_rtx_costs): Add correct costs for
        24-bit immediate add and 12-bit high immediate add.
        (TARGET_CONST_ANCHOR): Define.

gcc/testsuite/
        * gcc.target/aarch64/movk_3.c: New test.